My Replication of xyz

This replication study was part of my evaluation for the course Numerical Methods at SciencesPo Paris in Spring 2021

In this replication study, we do ...

Huggett.AssetGrid — Function

function AssetGrid(borrow)

Creates a grid of equally spaced points ranging from the minimum level of assets, i.e. the borrowing constraint level ‘borrow’ (which has to be specified by the user). The maximum number of assets ’a_max = 24’ and the number of assets ‘na=500’ have been specified for the sake of the replication exercise.

Inputs:

‘borrow::Float64’ : Borrowing constraint level for the Huggett economy. Notice that ‘borrow<=0’ must hold since in the Huggett economy considered the assets are not in positive net supply.

Outputs:

Column vector of dimension ‘na’ of type Matrix{Float64}

end